% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSpikeSlab-package.R
\docType{package}
\name{SequenceSpikeSlab-package}
\alias{SequenceSpikeSlab}
\alias{SequenceSpikeSlab-package}
\title{Fast Exact Bayesian Inference for the Sparse Normal Means Model}
\description{
The SequenceSpikeSlab package provides fast algorithms for exact Bayesian inference in the
sparse normal sequence model. It implements the methods of Van Erven and Szabo, 2021.
Special care has been taken to make the methods scale to large data sets, and to minimize
numerical errors (which arise in all software because floating point numbers are
represented with finite precision).
}
\details{
There are two main functions:
\code{\link{general_sequence_model}} and \code{\link{fast_spike_slab_beta}}.

For more details see the help vignette:
\code{vignette("SequenceSpikeSlab-vignette", package="SequenceSpikeSlab")}
}
\author{
\strong{Maintainer}: Tim van Erven \email{tim@timvanerven.nl}

Authors:
\itemize{
  \item Steven de Rooij \email{steven.de.rooij@gmail.com}
  \item Botond Szabo \email{botond.szabo@unibocconi.it}
}

}
\keyword{internal}
