% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSpikeSlab.R
\name{SSS_log_phi_psi_Laplace}
\alias{SSS_log_phi_psi_Laplace}
\title{Calculate log of phi and psi marginal densities for Laplace(lambda) slab}
\usage{
SSS_log_phi_psi_Laplace(x, sigma, lambda)
}
\arguments{
\item{x}{data vector}

\item{sigma}{standard deviation of observations}

\item{lambda}{parameter of Laplace slab density}
}
\value{
list (phi, psi), containing logs of phi and psi densities
}
\description{
Calculate log of densities phi and psi for data vector x, where
\deqn{phi[i] = Normal(x[i], sigma^2)}
\deqn{psi[i]) = E_Laplace(\theta)[Normal(x[i] - \theta, sigma^2)]}
}
