% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSpikeSlab.R
\name{SSS_make_beta_grid}
\alias{SSS_make_beta_grid}
\title{Creates a vector of uniformly spaced grid points in the beta parametrization
Ensures the number of generated grid points is >= mingridpoints (which does
not have to be integer), and that their number is always odd so there is always 
a grid point at pi/4.}
\usage{
SSS_make_beta_grid(minngridpoints)
}
\arguments{
\item{minngridpoints}{Minimum number of grid points}
}
\value{
Vector of betagrid points
}
\description{
Creates a vector of uniformly spaced grid points in the beta parametrization
Ensures the number of generated grid points is >= mingridpoints (which does
not have to be integer), and that their number is always odd so there is always 
a grid point at pi/4.
}
